Move odometry to `tankSubsystem`
With the new REV version, CANSparkMaxes with the same can ID cannot be created (it throws errors which kill the robot program)
