+++ src/main/java/frc/robot/subsystems/tank/TankSubsystem.java
@@ -4,10 +4,22 @@
 
 package frc.robot.subsystems.tank;
 
+import com.kauailabs.navx.frc.AHRS;
+
 import com.revrobotics.CANSparkMax;
 import com.revrobotics.CANSparkMax.IdleMode;
 import com.revrobotics.CANSparkMaxLowLevel.MotorType;
 
+import edu.wpi.first.math.MatBuilder;
+import edu.wpi.first.math.Nat;
+import edu.wpi.first.wpilibj.SPI;
+import edu.wpi.first.math.estimator.DifferentialDrivePoseEstimator;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.kinematics.DifferentialDriveWheelSpeeds;
+import edu.wpi.first.networktables.NetworkTableEntry;
+import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
+import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
 import static frc.robot.Constants.TankConstants.*;
@@ -19,12 +31,26 @@ public class TankSubsystem extends SubsystemBase {
   private final CANSparkMax rightMain;
   private final CANSparkMax rightFollow;
 
+  private final AHRS ahrs;
+  private final DifferentialDrivePoseEstimator poseEstimator;
+
+  private final ShuffleboardTab shuffleboardTab;
+  private final NetworkTableEntry shuffleboardXEntry;
+  private final NetworkTableEntry shuffleboardYEntry;
+  private final NetworkTableEntry shuffleboardAngleEntry;
+
+  public static final double ENCODER_ROTATIONS_TO_METERS = 5 / 92.08;
+
   public TankSubsystem() {
-    // Init left main and follower motors
+    // Init left main and follower motors and encoders
+    // Position conversion: Rotations -> m
+    // Velocity conversion: RPM -> m/s
     leftMain = new CANSparkMax(fLeftMotorPort, MotorType.kBrushless);
     leftMain.restoreFactoryDefaults();
     leftMain.setInverted(true);
     leftMain.setIdleMode(IdleMode.kBrake);
+    leftMain.getEncoder().setPositionConversionFactor(ENCODER_ROTATIONS_TO_METERS);
+    leftMain.getEncoder().setVelocityConversionFactor(ENCODER_ROTATIONS_TO_METERS / 60.0);
 
     leftFollow = new CANSparkMax(bLeftMotorPort, MotorType.kBrushless);
     leftFollow.restoreFactoryDefaults();
@@ -36,19 +62,37 @@ public class TankSubsystem extends SubsystemBase {
     rightMain.restoreFactoryDefaults();
     //rightMain.setInverted(true);
     rightMain.setIdleMode(IdleMode.kBrake);
+    rightMain.getEncoder().setPositionConversionFactor(ENCODER_ROTATIONS_TO_METERS);
+    rightMain.getEncoder().setVelocityConversionFactor(ENCODER_ROTATIONS_TO_METERS / 60.0);
 
     rightFollow = new CANSparkMax(bRightMotorPort, MotorType.kBrushless);
     rightFollow.restoreFactoryDefaults();
     rightFollow.follow(rightMain);
     rightFollow.setIdleMode(IdleMode.kBrake);
+
+    // Initialize navX AHRS
+    // https://www.kauailabs.com/public_files/navx-mxp/apidocs/java/com/kauailabs/navx/frc/AHRS.html
+    ahrs = new AHRS(SPI.Port.kMXP); 
+
+    // Initialize odometry class
+    // https://docs.wpilib.org/en/stable/docs/software/advanced-controls/state-space/state-space-pose_state-estimators.html
+    poseEstimator = new DifferentialDrivePoseEstimator(new Rotation2d(), new Pose2d(),
+      new MatBuilder<>(Nat.N5(), Nat.N1()).fill(0.02, 0.02, 0.01, 0.02, 0.02), // State measurement standard deviations. X, Y, theta.
+      new MatBuilder<>(Nat.N3(), Nat.N1()).fill(0.02, 0.02, 0.01), // Local measurement standard deviations. Left encoder, right encoder, gyro.
+      new MatBuilder<>(Nat.N3(), Nat.N1()).fill(0.1, 0.1, 0.01)); // Global measurement standard deviations. X, Y, and theta.
+
+    // Initialize Shuffleboard entries
+    shuffleboardTab = Shuffleboard.getTab("Drivetrain");
+    shuffleboardXEntry = shuffleboardTab.add("Robot x", 0).getEntry();
+    shuffleboardYEntry = shuffleboardTab.add("Robot y", 0).getEntry();
+    shuffleboardAngleEntry = shuffleboardTab.add("Gyro angle", 0).getEntry();
   }
 
   /**
    * Drive the system with the given power scales using the car system.
    * 
    * @param yScale       Scale in the forward/backward direction, from 1 to -1.
-   * @param angularScale Scale in the rotational direction, from 1 to -1,
-   *                     clockwise to counterclockwise.
+   * @param angularScale Scale in the rotational direction, from 1 to -1, clockwise to counterclockwise.
    */
   public void setCarDrivePowers(double yScale, double angularScale) {
     setCarDrivePowers(yScale, angularScale, true);
@@ -82,10 +126,8 @@ public class TankSubsystem extends SubsystemBase {
   /**
    * Drive the system with the given power scales using the tank system.
    * 
-   * @param leftScale  Scale in the forward/backward direction of the left motor,
-   *                   from -1 to 1.
-   * @param rightScale Scale in the forward/backward direction of the right motor,
-   *                   from -1 to 1.
+   * @param leftScale  Scale in the forward/backward direction of the left motor, from -1 to 1.
+   * @param rightScale Scale in the forward/backward direction of the right motor, from -1 to 1.
    */
   public void setTankDrivePowers(double leftScale, double rightScale, boolean squareInput) {
     if (squareInput) {
@@ -112,6 +154,73 @@ public class TankSubsystem extends SubsystemBase {
   public void setTankDriveVoltages(double leftVoltage, double rightVoltage) {
     leftMain.setVoltage(leftVoltage);
     rightMain.setVoltage(rightVoltage);
+  }
+
+  @Override
+  public void periodic() {
+    // Update odometry readings
+    Rotation2d gyroAngle = getRobotHeading();
+    DifferentialDriveWheelSpeeds wheelVelocities = getWheelSpeeds();
+    double leftDistance = leftMain.getEncoder().getPosition();
+    double rightDistance = rightMain.getEncoder().getPosition();
+
+    poseEstimator.update(gyroAngle, wheelVelocities, leftDistance, rightDistance);
+
+    // Update Shuffleboard entries
+    Pose2d pose = poseEstimator.getEstimatedPosition();
+    shuffleboardXEntry.setDouble(pose.getX());
+    shuffleboardYEntry.setDouble(pose.getY());
+    shuffleboardAngleEntry.setDouble(pose.getRotation().getDegrees());
+
+    System.out.println("Odometry readings: " + poseEstimator.getEstimatedPosition());
+    System.out.println("l: " + leftMain.getEncoder().getPosition());
+    System.out.println("r: " + rightMain.getEncoder().getPosition());
+  }
+
+    /**
+   * Gets the estimated current position of the robot.
+   * @return The estimated position of the robot as a Pose2d.
+   */
+  public Pose2d getRobotPosition() {
+    return poseEstimator.getEstimatedPosition();
+  }
+
+  /**
+   * Gets the gyro angle given by the NavX AHRS, inverted to be counterclockwise positive.
+   * @return The robot heading as a Rotation2d.
+   */
+  public Rotation2d getRobotHeading() {
+    return Rotation2d.fromDegrees(-ahrs.getAngle());
+  }
+
+  /**
+   * Gets the wheel speeds of the drivetrain.
+   * @return The drivetrain wheel speeds as a DifferentialDriveWheelSpeeds object.
+   */
+  public DifferentialDriveWheelSpeeds getWheelSpeeds() {
+    return new DifferentialDriveWheelSpeeds(
+      leftMain.getEncoder().getVelocity(), 
+      rightMain.getEncoder().getVelocity());
+  }
+
+  /**
+   * Reset the robot's position to a given Pose2d.
+   * @param position The position to reset the pose estimator to.
+   */
+  public void resetPosition(Pose2d position) {
+    leftMain.getEncoder().setPosition(0);
+    rightMain.getEncoder().setPosition(0);
+
+    // https://first.wpi.edu/wpilib/allwpilib/docs/release/java/edu/wpi/first/wpilibj/estimator/DifferentialDrivePoseEstimator.html#resetPosition(edu.wpi.first.wpilibj.geometry.Pose2d,edu.wpi.first.wpilibj.geometry.Rotation2d)
+    poseEstimator.resetPosition(position, getRobotHeading());
+  }
+
+  /**
+   * Zeros the robot's position.
+   * This method zeros both the robot's translation *and* rotation.
+   */
+  public void resetPosition() {
+    resetPosition(new Pose2d());
   }
 
   /**
